refactor: move subscribe logic to proposalactions
1. Move subscribe button to proposal actions
2. Sync local actions state to query on focus
